blerg
Former-commit-id: 07ac73c84dc31d2ff0e1b4cffd9d9f20b6828e6e

+++ CommonLib/Math/Vector/Vector2.h
@@ -162,7 +162,7 @@ namespace CommonUtilities
 		template <class TYPE>
 		const TYPE Length(const Vector2<TYPE>& aVector)
 		{
-			return sqrt((aVector.x * aVector.x) + (aVector.y*aVector.y));
+			return sqrtf((aVector.x * aVector.x) + (aVector.y*aVector.y));
 		};
 
 		template <class TYPE>
