+++ CommonLib/Math/Vector/Vector4.h
@@ -221,7 +221,7 @@ namespace CommonUtilities
 		template <typename T>
 		const T Length(const Vector4<T>& aVector)
 		{
-			return sqrt((aVector.x * aVector.x) +
+			return sqrtf((aVector.x * aVector.x) +
 				(aVector.y * aVector.y) +
 				(aVector.z * aVector.z) +
 				(aVector.w * aVector.w));
